added Settings
added a new text file where settings can be placed and then from around
the program values can be taken from the settings class. this way if we
want to change something we just have to change it once in the text
file. i'm going to go through now and change all of them.
ex. instead of 1600 for game width now we can use
Settings.getGameWidth()